requirements.txt
streamlit==1.24.1
pypdf2==3.0.1
langchain==0.0.228
python-dotenv==1.0.0
faiss-cpu==1.7.4
tiktoken==0.4.0
transformers
sentence-transformers
python-dotenv
llama-cpp-python==0.1.66
huggingface_hub
auto-gptq==0.2.2



